requirements.txt
python-magic==0.4.6
sqlalchemy
plyvel
requests
